Actualizacion servicios y token

+++ ValtikCA.Domain/Interfaces/IProductosXordenRepository.cs
@@ -5,7 +5,7 @@ namespace ValtikCA.Domain.Interfaces
     public interface IProductosXordenRepository
     {
         public IEnumerable<ProductosXorden> GetProductoXorden();
-        public ProductosXorden GetProductoXordenById(int Id);
+        public ProductosXorden GetProductoXordenById(decimal Id);
         public void InsertProductoXorden(ProductosXorden prodXOrden);
         public void UpdateProductoXordenById(ProductosXorden prodXOrden);
         public void DeleteProductoXordenById(ProductosXorden prodXOrden);
